requirements.txt
alembic==1.10.2
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.4.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
openai==0.8.0
packaging==23.0
password-strength==0.0.3.post2
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
